requirements.txt
albumentations==1.1.0
allennlp==1.0.0
astor==0.8.1
attrs==21.4.0
Babel==2.9.1
bce-python-sdk==0.8.64
beautifulsoup4==4.10.0
blis==0.4.1
boto3==1.21.27
botocore==1.24.27
cachetools==5.0.0
catalogue==1.0.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.0.4
cssselect==1.1.0
cssutils==2.4.0
cycler==0.11.0
cymem==2.0.6
Cython==0.29.28
decorator==5.1.1
diffimg==0.2.3
distlib==0.3.4
einops==0.2.0
et-xmlfile==1.1.0
filelock==3.0.12
flake8==4.0.1
Flask==2.0.3
Flask-Babel==2.0.0
fonttools==4.31.2
future==0.18.2
gast==0.3.3
gdown==4.4.0
h5py==3.6.0
identify==2.4.12
idna==3.3
imageio==2.16.1
imgaug==0.4.0
iniconfig==1.1.1
install==1.3.5
itsdangerous==2.1.2
jarowinkler==1.0.2
Jinja2==3.1.1
jmespath==1.0.0
joblib==1.1.0
jsonnet==0.18.0
jsonpickle==2.1.0
kiwisolver==1.4.1
Levenshtein==0.18.1
lmdb==1.0.0
lsd==0.3.3
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.1
mccabe==0.6.1
murmurhash==1.0.6
networkx==2.7.1
nltk==3.7
nodeenv==1.6.0
numpy==1.22.3
ocrd-fork-pylsd==0.0.3
opencv-contrib-python==4.4.0.46
opencv-python==4.1.2.30
opencv-python-headless==4.1.2.30
openpyxl==3.0.9
overrides==3.0.0
packaging==21.3
paddlepaddle==2.2.2
pandas==1.0.5
Pillow==9.0.1
plac==1.1.3
platformdirs==2.5.1
pluggy==1.0.0
pre-commit==2.17.0
prefetch-generator==1.0.1
premailer==3.10.0
preshed==3.0.6
protobuf==3.19.4
py==1.11.0
pyclipper==1.3.0.post2
pycodestyle==2.8.0
pycryptodome==3.14.1
pyflakes==2.4.0
pylsd==0.0.2
pyparsing==3.0.7
PySocks==1.7.1
pytest==7.1.1
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
rapidfuzz==2.0.7
regex==2022.3.15
requests==2.27.1
s3transfer==0.5.2
sacremoses==0.0.49
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
sentencepiece==0.1.96
Shapely==1.8.1.post1
shellcheck-py==0.8.0.4
six==1.16.0
soupsieve==2.3.1
spacy==2.2.4
srsly==1.0.5
tabulate==0.8.7
tensorboardX==2.5
thinc==7.4.0
threadpoolctl==3.1.0
tifffile==2022.3.25
timm==0.5.4
tokenizers==0.7.0
toml==0.10.2
tomli==2.0.1
torch==1.5.1
torchsummary==1.5.1
torchtext==0.6.0
torchvision==0.6.1
tqdm==4.47.0
transformers==2.11.0
trdg==1.7.0
typing-extensions==4.1.1
urllib3==1.26.9
virtualenv==20.14.0
visualdl==2.2.3
wasabi==0.9.0
Werkzeug==2.0.3
yacs==0.1.8
